# Faster parse_match (#440)