--- a/R/parse.R
+++ b/R/parse.R
@@ -79,7 +79,7 @@ parse_in_half <- function(x, char = "=") {
 
 parse_match <- function(x, pattern) {
   match_loc <- regexpr(pattern, x, perl = TRUE)
-  cap_start <- attr(match_loc,"capture.start")
+  cap_start <- attr(match_loc, "capture.start")
   cap_len <- attr(match_loc, "capture.length")
   cap_end <- (cap_start + cap_len - 1)
   cap_end[cap_end == -1] <- 0