--- a/R/parse.R
+++ b/R/parse.R
@@ -81,7 +81,7 @@ parse_match <- function(x, pattern) {
   match_loc <- regexpr(pattern, x, perl = TRUE)
   cap_start <- attr(match_loc, "capture.start")
   cap_len <- attr(match_loc, "capture.length")
-  cap_end <- (cap_start + cap_len - 1)
+  cap_end <- cap_start + cap_len - 1
   cap_end[cap_end == -1] <- 0
   pieces <- as.list(substring(x, cap_start, cap_end))
   pieces[pieces == ""] <- list(NULL)