@hadley 2024-02-10
For the future, can you please make sure to do `Fixes #12345` in the PR description? That'll make sure the related issue gets automatically closed when the PR is merged.